requirements.txt
dvc[s3,ssh]
biopython
tqdm
networkx
neo4j
docker
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyterlab
notebook
nbdime
click
nbresuse